refactor(find-wallet): inline detail bodies into their route files
WalletDetail and InterceptedWalletDetail each had exactly one caller whose
route file was a thin pass-through, so the indirection bought nothing. Move
each body into its route: the standalone layout into [wallet]/page.tsx and
the compact row-based one into @modal/(.)[wallet]/page.tsx.

Pure code motion — same JSX, same Matomo events. WalletDetailModal stays
separate as the client shell the modal route renders into.
